Factor out modal content rendering
Existing code used file graph nodes as input data to draw the QC modal. The new code uses QC objects instead, and the file graph modal code converts to that. Redmine #4828
